Prepare for PHP 7.4 preload
