+++ src/Symfony/Bundle/FrameworkBundle/DependencyInjection/CompatibilityServiceSubscriberInterface.contracts.php
@@ -1,0 +1,21 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\Bundle\FrameworkBundle\DependencyInjection;
+
+use Symfony\Contracts\Service\ServiceSubscriberInterface;
+
+/**
+ * @internal
+ */
+interface CompatibilityServiceSubscriberInterface extends ServiceSubscriberInterface
+{
+}
