+++ src/Symfony/Component/Cache/Exception/CacheException.php
@@ -11,15 +11,10 @@
 
 namespace Symfony\Component\Cache\Exception;
 
-use Psr\Cache\CacheException as Psr6CacheInterface;
 use Psr\SimpleCache\CacheException as SimpleCacheInterface;
 
 if (interface_exists(SimpleCacheInterface::class)) {
-    class CacheException extends \Exception implements Psr6CacheInterface, SimpleCacheInterface
-    {
-    }
+    require __DIR__.\DIRECTORY_SEPARATOR.'CacheException+psr16.php';
 } else {
-    class CacheException extends \Exception implements Psr6CacheInterface
-    {
-    }
+    require __DIR__.\DIRECTORY_SEPARATOR.'CacheException-psr16.php';
 }
