+++ src/Symfony/Component/Cache/Exception/InvalidArgumentException-psr16.php
@@ -11,10 +11,8 @@
 
 namespace Symfony\Component\Cache\Exception;
 
-use Psr\SimpleCache\LogicException as SimpleCacheInterface;
+use Psr\Cache\InvalidArgumentException as Psr6CacheInterface;
 
-if (interface_exists(SimpleCacheInterface::class)) {
-    require __DIR__.\DIRECTORY_SEPARATOR.'LogicException-psr16.php';
-} else {
-    require __DIR__.\DIRECTORY_SEPARATOR.'LogicException-psr16.php';
+class InvalidArgumentException extends \InvalidArgumentException implements Psr6CacheInterface
+{
 }
